Moving invocation methods to separate file
